update(driver/bpf/test): initial testing framework structure
